+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - FIXED: Panic on non-ELF object file.
 - FIXED: `const` string escapes no work correctly
+- FIXED: Report error on shift amounts greater than width
+- FIXED: Unify out-of-range shift behavior between const and layout phases
 - FIXED: Fixed double-backslash string escapes
 - INTERNAL: Added unit tests for mach-o, Windows PE and WASM object formats
 - README: Added region example to basic structure section.